Downgrade Kotlin to 1.6.x
It has come to our attention that when compiling the project with kotlin 1.9, projects using Kotlin 1.7.x and lower run into problems compiling.

This PR adjusts the Kotlin version to 1.6.21 (latest in 1.6.x) which should make it compatible with Kotlin 1.5.x, of which 1.5.0 was released in May 2021. This should provide sufficient backwards compatibility.