Introduce the hearbeat mechanism
Fixes #32. Extends the routes to include a bi-directional heartbeat route. Sending and receiving heartbeats should happen every 10 seconds. If heartbeats are missed for 30 seconds, the connection is set up from scratch again.

The server knows that we support heartbeats through the SetupPayload's feature flags. The API default to false for backwards compatibility.