Reduce logging to the minimum required for operation
The rest of the log lines are reduced to debug, so that users can debug their connection if needed. For normal operation, this is not needed.

From now on, we will log:
- Upon connection it will log it's connected and the settings retrieved
- upon the 5th failure of connection, it will log the connection error once

That's it. We don't want to spam the logs of the users when we deploy. The log during deployments is thus reduced to:

```
Connection to AxonIQ Console set up successfully! Settings: ClientSettingsV2(heartbeatInterval=10000, hear...
This application has lost it's connection to AxonIQ Console. Reconnection will be automatically attempted.
Connection to AxonIQ Console set up successfully! Settings: ClientSettingsV2(heartbeatInterval=10000, hear...
```

When we are offline, it will be similar to:
```
Connection to AxonIQ Console set up successfully! Settings: ClientSettingsV2(heartbeatInterval=10000, heartbeatTimeout=30000, processorReportInterval=2000, handlerReportInte
This application has lost it's connection to AxonIQ Console. Reconnection will be automatically attempted.
Failed to connect to AxonIQ Console. Error: io.netty.channel.AbstractChannel$AnnotatedConnectException: finishConnect(..) failed: Connection refused: localhost/127.0.0.1:700
```