Support decorators with null fields of type
A customer experienced the following error:
```
Failed to connect to AxonIQ Console. Error: null cannot be cast to non-null type T of io.axoniq.console.framework.UtilsKt.unwrapPossiblyDecoratedClass$lambda-0. Will keep trying to connect
```

This is because a field of an expected decorated type is null.
The code in this PR adjusts the method to find the first suitable non-null field. If none is found, it will return the decorator itself.