Allow users to configure instance identifiers
Right now, upon connection to AxonIQ Console, a random 4-letter string is assigned to make each client unique. Because the hostname can be the same for many clients (Especially in kubernetes). This works.

However, it's better if the clients themselves define this identifier randomly. As statistics are saved under the instance name, they will duplicate if an instance disconnects and reconnects. By moving this random generation to the client, we will solve this problem.