Add processing latency to the client
Measures the time between the currently processing message and the current clock. This will allow us to detect stalled segments, as this time will increase for every report.

Before, we only had ingest latency, but this will no longer show after 30 minutes. In addition, the nature of that metric does not allow it to increase. It will remain fixed during the processing.