Correct latency metrics in AxonIQ Console
First, we set the expiry time of ingest and commit latencies to two minutes, instead of 30 minutes. This prevents alarms from triggering and statistics to be strange.
I think 2 minutes is a balanced approach; generally batches will take under 10 seconds, so 2 minutes is already more than enough.

Second, we will now clean up the current message on cleanup, not on after commit. This will prevent hanging current latencies on errors.

Third, the commit latency is now the time between ingest of the first message of the batch and the commit time of it. As such, this represents the batch time.